+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 ### Changes
 - Updated translations: `bn_IN`, `am_ET`, `te_IN`, `pa_IN`, `hi_IN`, `bn_BD`
+- Added Registration Facility name and date in `PatientSummaryScreen`
 
 ### Fixes
 - Fixed occasional crash when starting a teleconsultation session ([LINK](https://app.clubhouse.io/simpledotorg/story/414/starting-a-teleconsultation-session-crashes-in-some-scenarios))